added code format

+++ src/events/client/SetupSystem.ts
@@ -1,7 +1,7 @@
-import { Event, Lavamusic } from '../../structures/index';
-import { Message, PermissionsBitField } from 'discord.js';
-import { oops, setupStart } from '../../utils/SetupSystem';
+import { Message, PermissionsBitField, TextChannel } from 'discord.js';
 
+import { Event, Lavamusic } from '../../structures/index';
+import { oops, setupStart } from '../../utils/SetupSystem';
 
 export default class SetupSystem extends Event {
     constructor(client: Lavamusic, file: string) {
@@ -10,30 +10,51 @@ export default class SetupSystem extends Event {
         });
     }
     public async run(message: Message): Promise<void> {
+        let channel = message.channel as any;
 
+        if (!(channel instanceof TextChannel)) return;
         if (!message.member.voice.channel) {
-            await oops(message.channel, `You are not connected to a voice channel to queue songs.`)
-            if (message) await message.delete().catch(() => { });
+            await oops(channel, `You are not connected to a voice channel to queue songs.`);
+            if (message) await message.delete().catch(() => {});
             return;
-        };
-
-        if (!message.member.voice.channel.permissionsFor(this.client.user).has(PermissionsBitField.resolve(['Connect', 'Speak']))) {
-            await oops(message.channel, `I don't have enough permission to connect/speak in ${message.member.voice.channel}`);
-            if (message) await message.delete().catch(() => { });
-            return;
-        };
-
-        if (message.guild.members.cache.get(this.client.user.id).voice.channel && message.guild.members.cache.get(this.client.user.id).voice.channelId !== message.member.voice.channelId) {
-            await oops(message.channel, `You are not connected to <#${message.guild.members.cache.get(this.client.user.id).voice.channelId}> to queue songs`);
-            if (message) await message.delete().catch(() => { });
-            return;
-        };
-        let player = this.client.queue.get(message.guildId)
-        if (!player) {
-            player = await this.client.queue.create(message.guild, message.member.voice.channel, message.channel, this.client.shoukaku.getNode());
         }
-        await setupStart(this.client, message.content, player, message)
-        if (message) await message.delete().catch(() => { });
+
+        if (
+            !message.member.voice.channel
+                .permissionsFor(this.client.user)
+                .has(PermissionsBitField.resolve(['Connect', 'Speak']))
+        ) {
+            await oops(channel,
+                `I don't have enough permission to connect/speak in <#${message.member.voice.channel.id}>`
+            );
+            if (message) await message.delete().catch(() => {});
+            return;
+        }
+
+        if (
+            message.guild.members.cache.get(this.client.user.id).voice.channel &&
+            message.guild.members.cache.get(this.client.user.id).voice.channelId !==
+                message.member.voice.channelId
+        ) {
+            await oops(channel,
+                `You are not connected to <#${
+                    message.guild.members.cache.get(this.client.user.id).voice.channelId
+                }> to queue songs`
+            );
+            if (message) await message.delete().catch(() => {});
+            return;
+        }
+        let player = this.client.queue.get(message.guildId);
+        if (!player) {
+            player = await this.client.queue.create(
+                message.guild,
+                message.member.voice.channel,
+                message.channel,
+                this.client.shoukaku.getNode()
+            );
+        }
+        await setupStart(this.client, message.content, player, message);
+        if (message) await message.delete().catch(() => {});
     }
 }
